[SAL] Introduce support for multiple memory types
Many architectures have different types of memory (flash, RAM, EEPROM,
etc), which is currently not well supported in the FAIL* toolchain.
Therefore, this change introduces memory types.

A memory type is a simple tag that is attached to memory events and
memory listeners. Furthermore, the generic-tracing experiment records
memory types and dump-trace displays them. Currently, they are not yet
used in importing. However, with SAIL integration, we require this
kind of disambiguation between different memory types.

Furthermore, MemoryAccessEvents now carry the memory area that was accessed.

This change should not break previous experiments and/or results.

Co-authored-by: Malte Bargholz <malte@screenri.de>