Use bit masks instead of access widths
With this *breaking* change, FAIL* does no longer use widths of data
accesses in its database schema, but bit masks. Thereby it is possible
to prevent or focus on the injection of individual bits.

For this change, the following database tables use masks with the
following meanings:

- trace.data_mask:    given bits were accessed
- fspgroup.data_mask: complete or partial match of a trace-table
                      entries data-mask
- fsppilot.data_mask: given bits are to be injected with this pilot

In order to reflect these database changes, you should
add (fspgroup.data_mask & trace.data_mask) to your join between trace
and fspgroup table to record for a cardinalty mismatch between both
tables. However, for the basic pruner (def-use pruner), this is not
necessary, as the data mask is in both cases 255.

Co-authored: Christian Dietrich <dietrich@sra.uni-hannover.de>