[SAIL] Integration of RISC-V and CHERI-RISCV
With this (large) patch, we integrate the riscv and cheri-riscv model
as potential tracing and injection targets. This change also
introduces the infratstructure to integrate other models with a
relatively low overhead.

Mixed into this patch are a few changes that are unrelated. I'm aware
of this, but I have no desire to pull everything appart. The effort
was large enough.

Co-authored-by: Christian Dietrich <dietrich@sra.uni-hannover.de>
Co-authored-by: Malte Bargholz <malte@screenri.de>
Co-authored-by: Marcel Budoj <marcel.budoj@googlemail.com>